refactor(ext/node): merge 02_init.js into 01_require.js
Drop the separate `02_init.js` module by inlining its content into
`01_require.js`. Along the way:

- Remove the `internals.requireImpl` bridge namespace — both files now
  live in the same module, so `setUsesLocalNodeModulesDir` collapses
  to a direct assignment and `nativeModuleExports` is a local
  reference. The dead `setInspectBrk` and `Module` keys go away too.
- Remove the unused `loadCjsModule` helper.
- Replace the nested `internals.node = { closeIdleConnections }`
  handle with a flat `internals.closeIdleConnections`, and update the
  test runner script string accordingly. `closeIdleConnections` is the
  only consumer of that namespace.
- Move the top-level `bindStreamsLazy(console, process)` call into
  `initialize`'s non-warmup branch. It only matters for `node:console`
  consumers, which can't run until after bootstrap anyway.
- Update `esm_entry_point` to `node:module` and drop `02_init.js`
  from the esm list.
