+++ ext/node/lib.rs
@@ -411,10 +411,9 @@ deno_core::extension!(deno_node,
     ops::http2::Http2Session,
     ops::http2::Http2Stream,
   ],
-  esm_entry_point = "ext:deno_node/02_init.js",
+  esm_entry_point = "node:module",
   esm = [
     dir "polyfills",
-    "02_init.js",
     "internal/streams/compose.js",
     "internal/streams/duplexpair.js",
     "internal/streams/lazy_transform.js",
